fixes for local sampling and additional options

+++ src/main/scala/au/org/ala/biocache/tool/SampleLocalRecords.scala
@@ -7,6 +7,7 @@ import au.org.ala.biocache.Config
 import au.org.ala.biocache.caches.LocationDAO
 import au.org.ala.biocache.index.BulkProcessor._
 import au.org.ala.biocache.util.{Json, OptionParser}
+import org.apache.commons.lang3.StringUtils
 import org.slf4j.LoggerFactory
 
 object SampleLocalRecords extends au.org.ala.biocache.cmd.Tool {
@@ -21,6 +22,8 @@ object SampleLocalRecords extends au.org.ala.biocache.cmd.Tool {
 
     var locFilePath = ""
     var keepFiles = false
+    var loadOccOnly = false
+    var sampleOnly = false
     var workingDir = Config.tmpWorkDir
     var batchSize = 100000
     var numThreads = 1
@@ -41,63 +44,72 @@ object SampleLocalRecords extends au.org.ala.biocache.cmd.Tool {
       intOpt("t", "threads", "The number of threads for the unique coordinate extract. The default is " + numThreads, {
         v: Int => numThreads = v
       })
+      opt("load-occ-only", "Just run the step that loads occ with values from loc", {
+        loadOccOnly = true
+      })
+      opt("sample-only", "Just run the step that samples and loads loc table", {
+        sampleOnly = true
+      })
     }
 
     if (parser.parse(args)) {
-      new SampleLocalRecords().sample(workingDir, numThreads, keepFiles)
+      new SampleLocalRecords().sample(workingDir, numThreads, keepFiles, loadOccOnly, sampleOnly)
     }
   }
 }
 
 class SampleLocalRecords {
 
-
-  def sample(workingDir:String, threads:Int, keepFiles:Boolean) : Unit = {
-
-    val locFilePath = workingDir + "/loc-local.txt"
-    val fw = new FileWriter(locFilePath)
-    val queue = new ConcurrentLinkedQueue[String]()
-
-    //export from loc table to file
-    Config.persistenceManager.pageOverLocal("loc", (key, map) => {
-      val lat = map.getOrElse("lat", "")
-      val lon = map.getOrElse("lon", "")
-      if(lat != "" && lon != ""){
-        queue.add(lon+"," + lat + "\n")
-      }
-      true
-    }, threads, Array("rowkey", "lat", "lon"))
-
-    val iter = queue.iterator()
-
-    while (iter.hasNext){
-      fw.write(iter.next())
-    }
-
-    fw.flush
-    fw.close
-
-    //run sampling
-    val sampling = new Sampling()
+  def sample(workingDir:String, threads:Int, keepFiles:Boolean, loadOccOnly:Boolean, sampleOnly:Boolean) : Unit = {
 
     val samplingFilePath = workingDir + "/sampling-local.txt"
+    val locFilePath = workingDir + "/loc-local.txt"
 
-    //generate sampling
-    sampling.sampling(locFilePath,
-      samplingFilePath,
-      singleLayerName="",
-      batchSize=100000,
-      concurrentLoading=true,
-      keepFiles=true
-    )
+    if(!loadOccOnly) {
 
-    //load sampling to occurrence records
-    logger.info("Loading sampling into occ table")
-    loadSamplingIntoOccurrences(threads)
-    logger.info("Completed loading sampling into occ table")
+      val fw = new FileWriter(locFilePath)
+      val queue = new ConcurrentLinkedQueue[String]()
+
+      //export from loc table to file
+      Config.persistenceManager.pageOverSelect("loc", (key, map) => {
+        val lat = map.getOrElse("lat", "")
+        val lon = map.getOrElse("lon", "")
+        if (lat != "" && lon != "") {
+          queue.add(lon + "," + lat + "\n")
+        }
+        true
+      }, "", "", 1000, "rowkey", "lat", "lon")
+
+      val iter = queue.iterator()
+
+      while (iter.hasNext) {
+        fw.write(iter.next())
+      }
+
+      fw.flush
+      fw.close
+
+      //run sampling
+      val sampling = new Sampling()
+      //generate sampling
+      sampling.sampling(locFilePath,
+        samplingFilePath,
+        singleLayerName = "",
+        batchSize = 100000,
+        concurrentLoading = true,
+        keepFiles = true
+      )
+    }
+
+    if(!sampleOnly) {
+      //load sampling to occurrence records
+      logger.info("Loading sampling into occ table")
+      loadSamplingIntoOccurrences(threads)
+      logger.info("Completed loading sampling into occ table")
+    }
 
     //clean up the file
-    if(!keepFiles){
+    if(!keepFiles && !loadOccOnly){
       logger.info(s"Removing temporary file: $samplingFilePath")
       (new File(samplingFilePath)).delete()
       if (new File(locFilePath).exists()) (new File(locFilePath)).delete()
@@ -107,9 +119,9 @@ class SampleLocalRecords {
   def loadSamplingIntoOccurrences(threads:Int) : Unit = {
     logger.info(s"Starting loading sampling for local records")
     Config.persistenceManager.pageOverLocal("occ", (guid, map) => {
-      val lat = map.getOrElse("decimalLatitude" + Config.persistenceManager.fieldDelimiter + "p", "")
-      val lon = map.getOrElse("decimalLongitude" + Config.persistenceManager.fieldDelimiter + "p", "")
-      if (lat != null && lon != null) {
+      val lat = map.getOrElse("decimallatitude" + Config.persistenceManager.fieldDelimiter + "p", "")
+      val lon = map.getOrElse("decimallongitude" + Config.persistenceManager.fieldDelimiter + "p", "")
+      if (lat != "" && lon != "") {
         val point = LocationDAO.getSamplesForLatLon(lat, lon)
         if (!point.isEmpty) {
           val (location, environmentalLayers, contextualLayers) = point.get
@@ -119,6 +131,8 @@ class SampleLocalRecords {
             false,
             false
           )
+        } else {
+          logger.info(s"[Loading sampling] Missing sampled values for $guid, with $lat, $lon")
         }
         counter += 1
         if (counter % 1000 == 0) {
@@ -126,6 +140,12 @@ class SampleLocalRecords {
         }
       }
       true
-    }, threads, Array("rowkey", "decimalLatitude" + Config.persistenceManager.fieldDelimiter + "p", "decimalLongitude" + Config.persistenceManager.fieldDelimiter + "p"))
+    }, threads,
+      Array(
+        "rowkey",
+        "decimalLatitude" + Config.persistenceManager.fieldDelimiter + "p",
+        "decimalLongitude" + Config.persistenceManager.fieldDelimiter + "p"
+      )
+    )
   }
 }
